requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
construct==2.5.3
defusedxml==0.5.0
Django==1.11.18
django-allauth==0.38.0
django-appconf==1.0.2
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-imagekit==4.0.2
django-mdeditor==0.1.9
django-reversion==3.0.2
docutils==0.14
future==0.17.1
httplib2==0.9.2
idna==2.8
imageio==2.4.1
jieba==0.39
Markdown==3.0.1
meditor==2.13.38
mtable==0.1.8
mysqlclient==1.4.1
numpy==1.16.0
oauthlib==3.0.0
pefile==2018.8.8
pilkit==2.0
Pillow==5.4.1
Pygments==2.3.1
PyMySQL==0.9.3
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-markdown-mathjax==1.0.4
python-ptrace==0.9.3
python3-openid==3.1.0
pytz==2018.9
QScintilla==2.10.8
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.7.2
urllib3==1.24.1
wcwidth==0.1.7
Whoosh==2.7.4
